cleanup: refactor live comments related functions to use-live-comments.js
